% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_classes.R
\name{balance_classes}
\alias{balance_classes}
\title{Balance training data classes}
\usage{
balance_classes(
  training_df,
  class_column,
  samples_per_class = NULL,
  seed_val = NULL
)
}
\arguments{
\item{training_df}{Data frame to be balanced using class column.}

\item{class_column}{Name of the column to balance the classes.}

\item{samples_per_class}{Optional. Number of samples per class. Defaults to
NULL and if NULL the samples_per_class will be the size of the smallest
class. If specify a samples_per_class > the smallest class size then
samples_per_class size will revert to smallest class size -
sample with replacement is never used.}

\item{seed_val}{Optional. Set seed_val to the same integer value
each time if want identical results returned.}
}
\value{
A data.frame with a subset of rows of input df after sampling.
}
\description{
Balance training data classes
}
\examples{
# Generate example df with unbalanced classes
df <- data.frame(
  ml_class = c(
    rep(0, 20),
    rep(1, 10),
    rep(2, 5)
  ),
  b1 = stats::runif(35, 0, 1)
)

# Show initial unbalanced classes
table(df$ml_class)

df <- balance_classes(df, "ml_class")

# Show now balanced and as samples_per_class not specified each class has
# smallest initial class size number of samples (5)

table(df$ml_class)

}
