% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_epsg.R
\name{check_epsg}
\alias{check_epsg}
\title{Check raster or vector objects have a valid crs epsg code
and if two inputs that they match.}
\usage{
check_epsg(x, y = NULL)
}
\arguments{
\item{x}{First raster or vector layer to check.}

\item{y}{Optional second raster or vector layer to compare.}
}
\value{
TRUE if epsg codes found and if they match when two inputs.
}
\description{
Check raster or vector objects have a valid crs epsg code
and if two inputs that they match.
}
\examples{
rast1 <- terra::rast(
  ncols = 10, nrows = 10, xmin = 0, xmax = 10, ymin = 0,
  ymax = 10, crs = "EPSG:4326"
)
rast2 <- terra::rast(
  ncols = 10, nrows = 10, xmin = 0, xmax = 10, ymin = 0,
  ymax = 10, crs = "EPSG:3857"
)

check_epsg(rast1, rast2)
}
