% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_with_box.R
\name{crop_with_box}
\alias{crop_with_box}
\title{Crop a raster using bounding box coords}
\usage{
crop_with_box(input_raster, xmin, ymin, xmax, ymax, out_file)
}
\arguments{
\item{input_raster}{\code{terra::rast} raster to be cropped.}

\item{xmin}{x min coordinate in same crs as \code{input_raster}}

\item{ymin}{y min coordinate in same crs as \code{input_raster}}

\item{xmax}{xmax min coordinate in same crs as \code{input_raster}}

\item{ymax}{xmax min coordinate in same crs as \code{input_raster}}

\item{out_file}{Full path and file name to write cropped image to.}
}
\value{
Cropped \code{terra::rast}
}
\description{
Crop a raster using bounding box coords
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
# create raster
rast <- terra::rast(
  ncols = 10, nrows = 10, xmin = 0, xmax = 10, ymin = 0,
  ymax = 10, crs = "EPSG:4326", vals = runif(100, 0, 1),
  names = c("b1")
)
# crop
crop_with_box(rast,
  xmin = 2,
  ymin = 2,
  xmax = 5,
  ymax = 5,
  out_file = "test.tif"
)
\dontshow{
setwd(.old_wd)
}

}
