% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_mask.R
\docType{data}
\name{example_mask}
\alias{example_mask}
\title{Example polygon feature to be as mask for use with sample raster
\code{example_s2()}.}
\format{
\subsection{\code{example_mask}}{

Simple feature collection with 1 feature and 0 fields.
Geometry type: POLYGON.
\describe{
\item{ml_class}{1 for forest and 0 for non-forest features}
}
}
}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/COPERNICUS_S2_SR}
}
\usage{
example_mask
}
\description{
Basic digitising of an example mask area over the sample Sentinel 2 image.
Manually captured from Sentinel 2 image for testing \code{rasterclassifier}
functions.\cr
}
\details{
Copernicus Sentinel data 2021.
Retrieved from ASF DAAC 15 April 2022, processed by ESA.
}
\keyword{datasets}
