% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_poly.R
\docType{data}
\name{example_poly}
\alias{example_poly}
\title{Example polygon features for use with sample raster \code{example_s2}.}
\format{
\subsection{\code{example_poly}}{

Simple feature collection with 27 features and 1 field.
Geometry type: POLYGON.
\describe{
\item{ml_class}{1 for forest and 0 for non-forest features}
}
}
}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/COPERNICUS_S2_SR}
}
\usage{
example_poly
}
\description{
Basic digitising of some forest and non-forest features for use as
training data for example code testing.\cr
Manually captured from Sentinel 2 image for testing \code{rasterclassifier}
functions.\cr
}
\details{
Copernicus Sentinel data 2021.
Retrieved from ASF DAAC 15 April 2022, processed by ESA.
}
\keyword{datasets}
