% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_pt.R
\docType{data}
\name{example_pt}
\alias{example_pt}
\title{Example point features for use with sample raster \code{example_s2}.}
\format{
\subsection{\code{example_pt}}{

Simple feature collection with 27 features and 1 field.
Geometry type: POINT.
\describe{
\item{ml_class}{1 for forest and 0 for non-forest features}
}
}
}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/COPERNICUS_S2_SR}
}
\usage{
example_pt
}
\description{
Point capture of some forest and non-forest for use as
training data for example code testing.\cr
Manually captured from Sentinel 2 image for testing \code{rasterclassifier}
functions.\cr
}
\details{
Copernicus Sentinel data 2021.
Retrieved from ASF DAAC 15 April 2022, processed by ESA.
}
\keyword{datasets}
