% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_s2.R
\docType{data}
\name{example_s2}
\alias{example_s2}
\title{Example Sentinel 2 image as \code{raster::brick()}}
\format{
\subsection{\code{example_s2}}{

A RasterBrick 147, 300, 44100, 5  (nrow, ncol, ncell, nlayers):
\describe{
\item{B2}{Blue}
\item{B3}{Green}
\item{B4}{Red}
\item{B8}{Near Infra Red}
\item{B11}{Short Wave Infra Red}
...
}
}
}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/COPERNICUS_S2_SR}
}
\usage{
example_s2
}
\description{
Small 3 km by 1.5 km Sentinel 2 image over Scotland\cr
Image date 21 April 2021
}
\details{
Copernicus Sentinel data 2021.
Retrieved from ASF DAAC 15 April 2022, processed by ESA.

Created with \code{raster::brick()} as at time of writing
\code{terra::rast()} SpatRaster
objects cannot be stored as rda files.
}
\keyword{datasets}
