% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pixels.R
\name{extract_pixels}
\alias{extract_pixels}
\title{Extract pixels from input raster using input training data.}
\usage{
extract_pixels(input_raster, training_shapes, training_class_col = "ml_class")
}
\arguments{
\item{input_raster}{terra::rast object for raster.}

\item{training_shapes}{sf spatial dataframe or terra::vect polygons for
training data.}

\item{training_class_col}{Optional column_name in training shapes showing
the class of each training data record. If not specified an "ml_class"
column is added with all values of 0.}
}
\value{
Dataframe of pixel values per class.
}
\description{
Extract pixels from input raster using input training data.
}
\examples{
# create raster
rast <- terra::rast(
ncols = 10, nrows = 10, xmin = 0, xmax = 10, ymin = 0,
ymax = 10, crs = "EPSG:4326", vals=runif(100, 0, 1),
names=c("b1")
)

# create polygon vector layer
p <- list(sf::st_polygon(list(cbind(c(0,1,1,0,0), c(0,0,1,1,0)))),
          sf::st_polygon(list(cbind(c(8, 9, 9, 8, 8), c(8, 8, 9, 9, 8)))))

vect <- sf::st_as_sf(data.frame(id=c(1,2),
geom=sf::st_sfc(p, crs="EPSG:4326")))

# extract raster pixels where intersect vector later
extract_pixels(rast, vect)
}
