% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_labels_select.R
\name{features_labels_select}
\alias{features_labels_select}
\title{Divide train and test dataframes to feature columns and label column.}
\usage{
features_labels_select(
  train_test_list,
  class_column = "ml_class",
  y_as_factor = TRUE
)
}
\arguments{
\item{train_test_list}{List of dataframes.}

\item{class_column}{Name of the column indicating \code{y} label values - this
column is dropped from features.}

\item{y_as_factor}{Should the label variable be converted to a
factor datatype? Defaults to TRUE as required for many R ml classification
algorithms.}
}
\value{
A list of  a list of feature \code{X} and label \code{y} dataframes,
for both train and test.
}
\description{
The list of train and test dataframes is split into sub lists of x and y.
}
\examples{
iris \%>\%
  train_test_split() \%>\%
  features_labels_select(class_column = "Species")
}
