% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mlc.R
\name{mlc}
\alias{mlc}
\title{Fit a maximum likelihood classifier using input lists
of train, test x, y.}
\usage{
mlc(train_test_xy)
}
\arguments{
\item{train_test_xy}{List of train and test lists, each sub-list containing
an X dataframe and a y vector.}
}
\value{
A custom object of class mlc
}
\description{
Designed to receive input of \code{df |> train_test_split() |> features_labels_select()}
}
\details{
A maximum likelihood classifier (mlc) is a traditional method used in remote
sensing. It assumes pixels in each band are normally distributed. The
mlc implemented in this package was written manually. Fitting an mlc stores
a vector of band means and a covariance matrix for each class. The
resulting mlc object can be passed to \code{\link[=predict]{predict()}}.
}
\examples{
input_list <- iris |>
  train_test_split() |>
  features_labels_select(class_column = "Species") |>
  mlc()
}
