% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classif_metrics.R
\name{get_classif_metrics}
\alias{get_classif_metrics}
\title{Calculate common classification metrics from prediction and actual vectors}
\usage{
get_classif_metrics(pred_y, actual_y)
}
\arguments{
\item{pred_y}{A vector of prediction class values. Should be factor
or integer.}

\item{actual_y}{A vector of actual class values. Should be the same datatype
and of the same length as \code{pred_y}.
.}
}
\value{
A list of metrics - confusion matrix, precision, recall, F1 Score
and overall accuracy.
}
\description{
Calculate common classification metrics from prediction and actual vectors
}
\examples{
actual_y <- iris$Species
pred_y <- sample(actual_y, length(actual_y), replace = TRUE)
get_classif_metrics(actual_y, pred_y)
}
