% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Split input dataframe into training and testing dataframes.}
\usage{
train_test_split(
  training_df,
  training_proportion = 0.75,
  shuffle = TRUE,
  seed_value = NULL
)
}
\arguments{
\item{training_df}{Input dataframe to be split.}

\item{training_proportion}{What proportion of nrows should be in training.
Defaults to 0.75.}

\item{shuffle}{Should the dataframe be shuffled before splitting. Defaults
to TRUE.}

\item{seed_value}{Seed value used in set.seed to ensure shuffling and
training testing splits are the same each time.}
}
\value{
list with a train and test dataframe as subsets of the input df.
}
\description{
Split input dataframe into training and testing dataframes.
}
\examples{
train_test_split(iris)
}
