% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_select.R
\name{xy_select}
\alias{xy_select}
\title{Divide train and test dataframes to x columns and y column.}
\usage{
xy_select(train_test_list, class_column, y_as_factor = TRUE)
}
\arguments{
\item{train_test_list}{List of dataframes.}

\item{class_column}{Name of the column indicating y values - this column is
dropped from x and is the only output of y.}

\item{y_as_factor}{Should the y variable by converted to a factor datatype.
Defaults to TRUE as required for many R ml classification algorithms.}
}
\value{
A list of  a list of x and y dataframes, for both train and test.
}
\description{
The list of train and test dataframes is split into sub lists of x and y.
}
