% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_train_test.R
\name{classif_train_test}
\alias{classif_train_test}
\title{Classifier testing n times against different train test splits.}
\usage{
classif_train_test(
  training_df,
  n_tests = 5,
  classifier = "random_forest",
  class_column = "ml_class",
  ...
)
}
\arguments{
\item{training_df}{Input dataframe to be used for training and testing
classifier.}

\item{n_tests}{How many different train-test splits and tests to run.}

\item{classifier}{Machine learning classifier to use. Currently the default
and only option is random_forest.}

\item{class_column}{Name of the column in training_df containing class
labels.}

\item{...}{Other arguments to be passed to classifier algorithm. For example
if random_forest can specify ntree for the number of decision trees used.}
}
\value{
List of metrics as determined by get_classif_metrics()
}
\description{
This process will print metrics specified by get_classif_metrics() and
a mean per metric over all tests.
}
