% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_random_forest.R
\name{fit_random_forest}
\alias{fit_random_forest}
\title{Fit a \code{\link[randomForest:randomForest]{randomForest::randomForest()}} using input lists
of train, test x, y.}
\usage{
fit_random_forest(train_test_xy, ...)
}
\arguments{
\item{train_test_xy}{List of train and test lists, each sub-list containing
an X dataframe and a y vector.}

\item{...}{Arguments other than x, y, testx, testy to be passed to
randomForest::randomForest(). For example might want to change ntree from
its default of 500.}
}
\value{
An object of class randomForest (see \code{\link[randomForest:randomForest]{randomForest::randomForest()}})
}
\description{
Designed to receive input of \verb{df |> train_test_split() |>\% features_labels_select()}
}
\examples{
input_list <- iris \%>\%
  train_test_split() \%>\%
  features_labels_select(class_column = "Species")
fit_random_forest(input_list, ntree = 100)
}
