% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mlc.R
\name{predict.mlc}
\alias{predict.mlc}
\title{Predict from a maximum likelihood classifier}
\usage{
\method{predict}{mlc}(object, newdata, ...)
}
\arguments{
\item{object}{a fitted mlc using \code{\link[=mlc]{mlc()}}}

\item{newdata}{a data.frame or SpatRaster of values to predict.}

\item{...}{other arguments (not relevant in this case)}
}
\value{
numeric vector of predicted class values or 1 band SpatRaster
of the same.
}
\description{
Designed to receive input of \code{df |> train_test_split() |> features_labels_select() |> fit_mlc()}
}
\details{
Can be used as a base \code{predict()} method or with \link[terra:predict]{terra::predict}.

The object from \code{\link[=mlc]{mlc()}} and new data to predict which can either
be data frame or SpatRaster object. Needs to have the same bands and band
order as used in fitting the mlc. If input is a SpatRaster returns
SpatRaster, otherwise returns vector of class predictions.
}
\examples{
\dontrun{
terra::predict(example_img, model = mlc, na.rm = TRUE)
}
}
